Array and Field synchronization and array unit tests.
Set up path-forward to allow Field to be natural-ordered or fortran-ordered,
currently hard-wired to be fortran-ordered.

+++ packages/phdmesh/include/mesh/FieldParallel.hpp
@@ -56,7 +56,7 @@ bool communicate_field_data(
 void communicate_field_data(
   ParallelMachine ,
   const std::vector<EntityProc> & ,
-  const size_t field_count ,
+  const unsigned field_count ,
   const FieldBase * fields[] ,
   CommAll & sparse );
